requirements.txt
web3==6.11.3
python-dotenv==1.0.0
